Improve reconstructor (#59)
- Update dependency versions
- Refactor to simplify plugins
- Use floats for scan positions
- Display message dialog when attempting to edit item without parameters
- Include initializer name in probe/object table views
- Handle exceptions when file saving fails
- Support saving multi-mode probes to CSV
- Transpose array dimensions when reading/writing MATLAB files
- Add reader for Bionanoprobe MDA scan files
- Support power decay and orthogonalization when initializing additional probe modes
- Change scan/object mapping: midpoint of scan bounding box now anchored to center of object array
- Object padding is customizable; object grows to maintain center point
- Refactor state data input/output: output now includes object center
- Refactor model to reduce code duplication between reconstructor library adapters
- Reconstruction results are only added to the scan/probe/object repositories when changed
- Add support for extracting object patches via interpolation
- Add controls for changing tike log level
- PtychoNN training and inference are now working
- Implement circular buffers to support PtychoNN continual training workflows
- Can save PtychoNN training data to NPZ file
- Display loss plot in reconstructor view after training
- Add batch mode training hook for workflows
- Add reconstructor API for workflows

+++ ptychodus/model/scan/sizer.py
@@ -1,5 +1,4 @@
 from __future__ import annotations
-from decimal import Decimal
 from typing import Optional
 
 from ...api.geometry import Box2D, Interval
@@ -22,19 +21,19 @@ class ScanSizer(Observable, Observer):
         scan.addObserver(sizer)
         return sizer
 
-    def getBoundingBoxInMeters(self) -> Optional[Box2D[Decimal]]:
-        boundingBoxInMeters: Optional[Box2D[Decimal]] = None
+    def getBoundingBoxInMeters(self) -> Optional[Box2D[float]]:
+        boundingBoxInMeters: Optional[Box2D[float]] = None
 
         if self._settings.expandBoundingBox.value:
-            rangeXInMeters = Interval[Decimal](
-                self._settings.boundingBoxMinimumXInMeters.value,
-                self._settings.boundingBoxMaximumXInMeters.value,
+            rangeXInMeters = Interval[float](
+                float(self._settings.boundingBoxMinimumXInMeters.value),
+                float(self._settings.boundingBoxMaximumXInMeters.value),
             )
-            rangeYInMeters = Interval[Decimal](
-                self._settings.boundingBoxMinimumYInMeters.value,
-                self._settings.boundingBoxMaximumYInMeters.value,
+            rangeYInMeters = Interval[float](
+                float(self._settings.boundingBoxMinimumYInMeters.value),
+                float(self._settings.boundingBoxMaximumYInMeters.value),
             )
-            boundingBoxInMeters = Box2D[Decimal](rangeXInMeters, rangeYInMeters)
+            boundingBoxInMeters = Box2D[float](rangeXInMeters, rangeYInMeters)
 
         selectedScan = self._scan.getSelectedItem()
 
@@ -49,15 +48,16 @@ class ScanSizer(Observable, Observer):
             return boundingBoxInMeters
 
         if boundingBoxInMeters is None:
-            rangeXInMeters = Interval[Decimal](point.x, point.x)
-            rangeYInMeters = Interval[Decimal](point.y, point.y)
-            boundingBoxInMeters = Box2D[Decimal](rangeXInMeters, rangeYInMeters)
+            rangeXInMeters = Interval[float](point.x, point.x)
+            rangeYInMeters = Interval[float](point.y, point.y)
+            boundingBoxInMeters = Box2D[float](rangeXInMeters, rangeYInMeters)
         else:
             boundingBoxInMeters = boundingBoxInMeters.hull(point.x, point.y)
 
         for point in it:
             boundingBoxInMeters = boundingBoxInMeters.hull(point.x, point.y)
 
+        # TODO cache this
         return boundingBoxInMeters
 
     def update(self, observable: Observable) -> None:
